@@ -7,31 +7,33 @@
   "type": "module",
   "scripts": {
     "tsc": "tsc",
-    "build": "rm -rf dist && tsc -b tsconfig.build.json",
+    "build:esm": "tsc -b && echo '{\"type\": \"module\"}' > dist/esm/package.json",
+    "build:cjs": "tsc -b tsconfig.cjs.json && echo '{\"type\": \"commonjs\"}' > dist/cjs/package.json",
+    "build": "rm -rf dist && yarn build:esm && yarn build:cjs",
     "release": "npm run build && npm test",
     "ncu": "ncu",
     "test": "echo 'No tests'"
   },
   "exports": {
     "./*": {
-      "import": "./dist/*.js",
-      "require": "./dist/*.js"
+      "import": "./dist/esm/*.js",
+      "require": "./dist/cjs/*.js"
     },
     "./*.d.ts": {
-      "import": "./dist/*.d.ts",
-      "require": "./dist/*.d.ts"
+      "import": "./dist/esm/*.d.ts",
+      "require": "./dist/cjs/*.d.ts"
     }
   },
   "peerDependencies": {
     "@vramework/core": "^0.4.1"
   },
   "dependencies": {
-    "@types/aws-lambda": "^8.10.136",
+    "@types/aws-lambda": "^8.10.145",
     "aws-lambda": "^1.0.7",
-    "cookie": "^1.0.0",
-    "path-to-regexp": "^8.1.0"
+    "cookie": "^1.0.1",
+    "path-to-regexp": "^8.2.0"
   },
   "devDependencies": {
-    "typescript": "^5.4.2"
+    "typescript": "^5.6.3"
   }
 }
