@@ -7,19 +7,21 @@
   "type": "module",
   "scripts": {
     "tsc": "tsc",
-    "build": "rm -rf dist && tsc -b tsconfig.build.json",
+    "build:esm": "tsc -b && echo '{\"type\": \"module\"}' > dist/esm/package.json",
+    "build:cjs": "tsc -b tsconfig.cjs.json && echo '{\"type\": \"commonjs\"}' > dist/cjs/package.json",
+    "build": "rm -rf dist && yarn build:esm && yarn build:cjs",
     "release": "npm run build && npm test",
     "ncu": "ncu",
     "test": "echo 'No tests'"
   },
   "exports": {
     "./*": {
-      "import": "./dist/*.js",
-      "require": "./dist/*.js"
+      "import": "./dist/esm/*.js",
+      "require": "./dist/cjs/*.js"
     },
     "./*.d.ts": {
-      "import": "./dist/*.d.ts",
-      "require": "./dist/*.d.ts"
+      "import": "./dist/esm/*.d.ts",
+      "require": "./dist/cjs/*.d.ts"
     }
   },
   "peerDependencies": {
@@ -28,11 +30,11 @@
   "dependencies": {
     "@types/cookie": "^0.6.0",
     "@vramework/core": "^0.4.1",
-    "cookie": "^1.0.0",
+    "cookie": "^1.0.1",
     "eventemitter3": "^5.0.1"
   },
   "devDependencies": {
-    "next": "^14.2.9",
-    "typescript": "^5.4.2"
+    "next": "^14.2.15",
+    "typescript": "^5.6.3"
   }
 }
