@@ -8,19 +8,21 @@
   "type": "module",
   "scripts": {
     "tsc": "tsc",
-    "build": "rm -rf dist && tsc -b tsconfig.build.json",
+    "build:esm": "tsc -b && echo '{\"type\": \"module\"}' > dist/esm/package.json",
+    "build:cjs": "tsc -b tsconfig.cjs.json && echo '{\"type\": \"commonjs\"}' > dist/cjs/package.json",
+    "build": "rm -rf dist && yarn build:esm && yarn build:cjs",
     "release": "npm run build && npm test",
     "ncu": "ncu",
     "test": "echo 'No tests'"
   },
   "exports": {
     "./*": {
-      "import": "./dist/*.js",
-      "require": "./dist/*.js"
+      "import": "./dist/esm/*.js",
+      "require": "./dist/cjs/*.js"
     },
     "./*.d.ts": {
-      "import": "./dist/*.d.ts",
-      "require": "./dist/*.d.ts"
+      "import": "./dist/esm/*.d.ts",
+      "require": "./dist/cjs/*.d.ts"
     }
   },
   "peerDependencies": {
@@ -29,10 +31,10 @@
   "dependencies": {
     "@types/qs": "^6",
     "qs": "^6.13.0",
-    "query-string": "^9.1.0",
-    "uWebSockets.js": "uNetworking/uWebSockets.js#v20.48.0"
+    "query-string": "^9.1.1",
+    "uWebSockets.js": "uNetworking/uWebSockets.js#v20.49.0"
   },
   "devDependencies": {
-    "typescript": "^5.4.2"
+    "typescript": "^5.6.3"
   }
 }
