@@ -1,6 +1,7 @@
 import { VrameworkRequest } from '@vramework/core/vramework-request'
 import { Request } from 'express-serve-static-core'
-import getRawBody from 'raw-body'
+import * as getRawBodyImp from 'raw-body'
+const getRawBody = 'default' in getRawBodyImp ? (getRawBodyImp.default as any) : getRawBodyImp
 
 export class VrameworkExpressRequest extends VrameworkRequest {
   constructor(private request: Request) {
