@@ -1,6 +1,6 @@
 import { JSONValue } from '@vramework/core/types/core.types'
 import { VrameworkResponse } from '@vramework/core/vramework-response'
-import cookie from 'cookie'
+import * as cookie from 'cookie'
 import { FastifyReply } from 'fastify'
 
 export class VrameworkFastifyResponse extends VrameworkResponse {
