@@ -1,6 +1,6 @@
 import { HttpRequest, HttpResponse } from 'uWebSockets.js'
-import cookie from 'cookie'
-import querystring from 'qs'
+import * as cookie from 'cookie'
+import * as querystring from 'qs'
 import { VrameworkRequest } from '@vramework/core/vramework-request'
 
 export class VrameworkUWSRequest extends VrameworkRequest {
