@@ -1,5 +1,5 @@
 import { HttpResponse } from 'uWebSockets.js'
-import cookie from 'cookie'
+import * as cookie from 'cookie'
 import { VrameworkResponse } from '@vramework/core/vramework-response'
 import { JSONValue } from '@vramework/core/types/core.types'
 
